% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{add_global_mcmc}
\alias{add_global_mcmc}
\title{Add an MCMC chain to a global run of FPEM}
\usage{
add_global_mcmc(
  run_name,
  output_dir_path = file.path(getwd(), "output"),
  chain_nums = 2,
  run_in_parallel = isTRUE(length(chain_nums) > 1),
  verbose = deprecated(),
  output_folder_path = deprecated()
)
}
\arguments{
\item{run_name}{The name of the run to add a chain to. Unlike
\code{\link{do_global_mcmc}}, this is must be specified.}

\item{output_dir_path}{Path of the top-level directory in which to save
results. Sub-directories named \code{run_name} will be created inside
this directory for each run.}

\item{chain_nums}{number identifying chains to add. Unlike
\code{\link{do_global_mcmc}}, this can be a scalar. It must \emph{not}
be the number identifying a chain already created. See
\dQuote{Description}.}

\item{run_in_parallel}{Logical. Determines if MCMC chains are run in
parallel. Parallel running requires package
#ifdef windows
\pkg{doParallel}.
#endif
#ifndef windows

\pkg{doParallel} or \pkg{doMC}.
#endif
Defaults to serial running if \code{run_in_parallel = TRUE} but the package
is not available.}
}
\value{
A name for the run is returned invisibly as a character string. MCMC
chains are saved to \file{run_dir_path/temp.JAGSobjects}. They need to
be post-processed with \code{\link{post_process_mcmc}}. The run name
must be passed to \code{\link{post_process_mcmc}} to locate find the
saved chains for processing. Run names for married and unmarried runs
must also be passed to \code{\link{combine_runs}} to generate all women
MCMC results.
}
\description{
At least two chains are required to post-process the results of a run of the
FPEM model. Multiple chains can be run by \code{\link{do_global_mcmc}} but, if additional chains are desired,
this function can be used to add them. There must be at least one chain
already started (identified via \code{run_name}).
}
\examples{
vignette("FPEMglobal_Intro")
}
\author{
Mark Wheldon, Andrew Tait
}
