% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{add_special_aggregates}
\alias{add_special_aggregates}
\title{Add special aggregates to a global run of FPEM}
\usage{
add_special_aggregates(
  output_folder_path,
  countries_for_aggregates_csv_filename = NULL,
  denominator_counts_csv_filename = NULL,
  start_year = NULL,
  end_year = NULL,
  years_change = NULL,
  years_change2 = NULL,
  model_diagnostics = FALSE,
  special_aggregates_name,
  verbose = FALSE
)
}
\description{
This is a convenience function to add (possibly more) special
aggregates to a completed global run. It wraps calls to
\code{\link{post_process_mcmc}} and \code{make_results}.
}
\author{
Mark Wheldon
}
