% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregates_wfr2024_locations}
\alias{aggregates_wfr2024_locations}
\title{Country and Area Aggregations from World Fertility Report 2024}
\format{
\preformatted{'data.frame':	237 obs. of  6 variables:
$ LocID          : num  108 174 262 232 231 404 450 454 480 175 ...
$ LocTypeID      : int  4 4 4 4 4 4 4 4 4 4 ...
$ LocTypeName    : chr  "Country/Area" "Country/Area" "Country/Area" "Country/Area" ...
$ LocationID     : num  108 174 262 232 231 404 450 454 480 175 ...
$ ParentPrintName: chr  "After 2054" "After 2054" "1994-2054" "After 2054" ...
}
\source{
United Nations, Department of Economic and Social Affairs, Population Division
}
\usage{
aggregates_wfr2024_locations
}
\description{
A data frame containing the aggregations of countries and areas
used in United Nations' \emph{World Fertility Report} (\acronym{WFR}), 2024
edition. These are defined by the year in which the total
fertility rate (\acronym{TFR}) reached 2.1, births per woman, or
is expected to reach 2.1 based on the medium projection scenario
in World Population Prospects (\acronym{WPP}), 2024 edition:
}
\details{
\describe{
\item{Before 1994}{Countries reaching low fertility before 1994}
\item{1994-2054}{Countries reaching low fertility between 1994 and 2054}
\item{After 2054}{Countries potentially reaching low fertility later in the century after 2054}
}

\code{aggregates_wfr2024_locations} is a data frame containing the
single aggregation list for this set. The format is similar to the
\dQuote{Locations} elements of the lists in
\code{\link{aggregates_wpp2024_list}}.
}
\references{
United Nations. (in prep.). \emph{World Fertility Report 2024}. New York: United
Nations Department of Economic and Social Affairs, Population
Division.
}
\seealso{
Other special aggregations: 
\code{\link{aggregates_wpp2024_list}}
}
\concept{special aggregations}
\keyword{datasets}
