% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{assert_valid_output_dir}
\alias{assert_valid_output_dir}
\title{Check that a directory is a valid FPEMglobal output directory}
\usage{
assert_valid_output_dir(
  output_folder_path,
  post_processed = TRUE,
  countrytrajectories = post_processed,
  made_results = post_processed,
  adjusted_medians = post_processed,
  age_ratios = NULL,
  verbose = FALSE
)
}
\arguments{
\item{output_folder_path}{Path to directory to validate.}

\item{post_processed}{Logical; has \code{\link{post_process_mcmc}}
been run on the directory?}

\item{countrytrajectories}{Logical; check for
\file{countrytrajectories} or \file{aggregatetrajectories}
directories?}

\item{made_results}{Logical; has \code{\link{make_results}} been
run on the directory?}

\item{adjusted_medians}{Logical; check for adjusted median
results?}

\item{age_ratios}{Logical; check for age ratio results? Defaults
to \code{FALSE} if \code{output_folder_path} points to a
\dQuote{15-49} run, otherwise the value of
\code{post_processed}.}
}
\value{
If all checks pass, \code{output_folder_path} is returned
invisibly, otherwise an error is thrown.
}
\description{
Checks the content of \code{output_dir} to make sure certain
directories and files are present. If some are missing, an error
is returned, otherwise \code{output_folder_path} is returned
invisibly.
}
\details{
The default for 'post_processed' is 'TRUE' by because an
un-processed directory doesn't even have 'mcmc.array.rda', which
means it's unlikely to be used.
}
\author{
Mark Wheldon
}
