% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{combine_runs}
\alias{combine_runs}
\title{Make all women results from married and unmarried women FPEM runs}
\usage{
combine_runs(
  run_desc = "",
  married_women_run_name,
  married_women_run_output_folder_path = NULL,
  unmarried_women_run_name,
  unmarried_women_run_output_folder_path = NULL,
 
    unmarried_women_run_data_folder_path = file.path(unmarried_women_run_output_folder_path,
    "data"),
  region_information_csv_filename = "country_and_area_classification.csv",
  special_aggregates_name = NULL,
  denominator_counts_csv_filename = NULL,
  countries_for_aggregates_csv_filename = "countries_mwra_195.csv",
  countries_in_CI_plots_csv_filename = "countries_mwra_195.csv",
  output_folder_path = NULL,
  start_year = 1970.5,
  end_year = 2030.5,
  years_change = matrix(c(1990.5, 2000.5, 2000.5, 2019.5, 2019.5, 2030.5, 2012.5,
    2019.5, 2012.5, 2019.5, 2012.5, 2020.5), ncol = 2, byrow = TRUE),
  years_change2 = matrix(c(2005.5, 2010.5, 2015.5, 2000.5, 2005.5, 2010.5, 1995.5,
    2000.5, 2005.5, 1990.5, 1995.5, 2000.5, 1990.5, 2000.5, 2010.5, 2000.5, 2010.5,
    2019.5), ncol = 3, byrow = TRUE),
  make_any_aggregates = TRUE,
  adjust_medians = FALSE,
  age_ratios_age_total_unmarried_run_name = NULL,
  age_ratios_age_total_married_run_name = NULL,
  age_ratios_age_total_all_women_run_name = NULL,
  age_ratios_age_total_unmarried_output_folder_path = NULL,
  age_ratios_age_total_married_output_folder_path = NULL,
  age_ratios_age_total_all_women_output_folder_path = NULL,
  age_ratios_age_total_denominator_counts_csv_filename = "number_of_women_15-49.csv",
  age_ratios_age_total_denominator_counts_folder_path = NULL,
  run_name_override = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_desc}{Character. Brief note to be appended to the
auto-generated \code{run_name}. Ignored if
\code{run_name_override} is non-\code{NULL}.}

\item{married_women_run_name}{Run name of the married women run to
be combined. Ignored if
\code{married_women_run_output_folder_path} is
\code{NULL}. Otherwise, the output folder will be assumed to
be \code{file.path("output", married_women_run_name)}.}

\item{married_women_run_output_folder_path}{Path to the folder
containing results for the married women run to be combined.}

\item{unmarried_women_run_name}{Same as
\code{married_women_run_name} but for the unmarried women run.}

\item{unmarried_women_run_output_folder_path}{Path to the folder
containing results for the unmarried women run to be combined.}

\item{unmarried_women_run_data_folder_path}{Path to the folder
containing results for the unmarried women run to be
combined. (Only used if \code{special_aggregates_name} is
non-\code{NULL}.}

\item{region_information_csv_filename}{Filename of the \file{.csv}
file containing classifications of countries in sub-regions,
regions, etc. See \dQuote{Details}.}

\item{special_aggregates_name}{Character vector of names
(\emph{not} filenames) of any speical aggregates
desired. There must be a corresponding file with name
\file{\code{special_aggregates_name}.csv} in
\code{input_data_folder_path} that defines the special
aggregates. See \dQuote{Details}.}

\item{denominator_counts_csv_filename}{Name of the \file{.csv}
file containing estimates and projections of the number of
women by marital status, age, and year. See \dQuote{Details}.}

\item{countries_for_aggregates_csv_filename}{Name of the
\file{.csv} file listing countries that will be used in constructing country aggregates.}

\item{countries_in_CI_plots_csv_filename}{Name of \file{.csv} file
that lists the countries to be included in the main
country-level indicator plots. See \code{\link{make_results}}
for further details. \code{combine_runs} will copy this file
to \file{\code{output_folder_path}/data} so it is available
for \code{make_results}.}

\item{output_folder_path}{Filepath to directory where outputs
should be saved. If \code{NULL}, defaults to
\code{file.path("output", run_name)}.}

\item{start_year}{Estimates and projections are produced for a
specified time interval. This is the start year of that
interval.}

\item{end_year}{Estimates and projections are produced for a
specified time interval. This is the end year of that
interval.}

\item{years_change}{A two-column matrix giving the year pairs (as
rows) between which probabilistic estimates of changes in the
indicators are desired.}

\item{years_change2}{A three-column matrix giving the year triples
(as rows) among which to compute probabilistic estimates of
change-in-changes.}

\item{make_any_aggregates}{Logical. Should country aggregates of
any kind (including default aggregates) be produced?}

\item{adjust_medians}{Logical. Should adjusted medians outputs be
produced in addition to unadjusted outputs?}

\item{age_ratios_age_total_unmarried_run_name}{Run name of the
unmarried 15--49 run to use as the denominator for age ratios.}

\item{age_ratios_age_total_married_run_name}{Run name of the
married 15--49 run to use as the denominator for age ratios.}

\item{age_ratios_age_total_all_women_run_name}{Run name of the all
women 15--49 run to use as the denominator for age ratios.}

\item{age_ratios_age_total_unmarried_output_folder_path}{File path
to output directory of the unmarried 15--49 run to use to make
age ratios.}

\item{age_ratios_age_total_married_output_folder_path}{File path
to output directory of the married 15--49 run to use to make
age ratios.}

\item{age_ratios_age_total_all_women_output_folder_path}{File path
to output directory of the all women 15--49 run to use to make
age ratios.}

\item{age_ratios_age_total_denominator_counts_csv_filename}{Name of
the \file{.csv} file containing estimates and projections of
the number of women by marital status, age, and year, for the
age group 15--49. Only used if \code{make_age_ratios} is
\code{TRUE}. Searched for in \code{age_ratios_age_total_denominator_counts_folder_path}.}

\item{age_ratios_age_total_denominator_counts_folder_path}{Path to
\code{age_ratios_age_total_denominator_counts_csv_filename}. If
\code{NULL}, defaults to
\code{file.path(age_ratios_age_total_output_folder_path, "data")}.}

\item{run_name_override}{Character. User defined run name to
override default generation. \code{run_desc} is ignored if
this is non-\code{NULL}.}

\item{verbose}{Logical; print lots and lots of messages about
progress?}
}
\value{
A name for the run returned invisibly as a character
string. Results are saved to \file{\code{output_folder_path}}.
}
\description{
Combines MCMC outputs from married and unmarried women FPEM runs
to make MCMC outputs for all women. The MCMC outputs must have
been produced by \code{\link{do_global_mcmc}} or
\code{\link{do_global_run}} (the former is sufficient).
}
\details{
Married \emph{and} unmarried women runs must have been produced
previously by, e.g,. \code{\link{do_global_run}}. They are
identified by their output directory filepaths
(\code{married_women_run_output_folder_path} and
\code{unmarried_women_run_output_folder_path}). Files from the
married and unmarried runs that are needed to make all women
results are copied to \code{output_folder_path} (and
sub-directories) and noted in files called \file{log.txt} in the
respective directories. The file
\file{\code{output_folder_path}/log.txt} provides a record of
which married and unmarried women runs were combined.
}
\examples{
vignette("FPEMglobal_Intro").

}
\seealso{
\code{\link{do_global_all_women_run}} to do married, unmarried,
\emph{and all women runs}, and produce results, all in one
call.
}
\author{
Mark Wheldon, Andrew Tait.
}
