% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{compare_runs_CI_plots}
\alias{compare_runs_CI_plots}
\title{Compare two global runs by plotting}
\usage{
compare_runs_CI_plots(
  run_name_1,
  run_name_2,
  run_1_output_folder_path = file.path("output", run_name_1),
  run_1_plot_label = run_name_1,
  run_2_output_folder_path = file.path("output", run_name_2),
  run_2_plot_label = run_name_2,
  output_folder_path = file.path(run_1_output_folder_path, "fig", "compare_runs_plots"),
  plot_data = NULL,
  all_women = NULL,
  compare_aggregates = TRUE
)
}
\arguments{
\item{run_1_output_folder_path}{Filepath to directory where
outputs for \code{run_1_name} are saved.}

\item{run_1_plot_label}{The label to use for \dQuote{run_1} in the
plots. Defaults to \code{run_1_name}.}

\item{run_2_output_folder_path}{See
\code{run_1_output_folder_path}, but for \dQuote{run_2}.}

\item{run_2_plot_label}{See \code{run_2_plot_label}, but for
\dQuote{run_2}.}

\item{output_folder_path}{Filepath to the directory in which the
comparison plots will be saved.}

\item{plot_data}{Logical; should available input data be plotted?
Defaults to \code{TRUE} unless either \code{all_women} is
\code{TRUE} or \dQuote{run_1} and \dQuote{run_2} are from
different marital groups.}

\item{all_women}{Logical; is the run an all women run such as the
kind produced by \code{\link{combine_runs}} or
\code{\link{do_global_all_women_run}}? If \code{NULL} an
attempt is made to determine this automatically from
\file{mcmc.meta.rda} in the output folder of
\dQuote{run_1}. Note that either both or neither must be
either all women runs.}

\item{compare_aggregates}{Logical; should comparison plots be made
for the aggregates as well as the countries? If \code{FALSE}
only plots for countries are produced.}

\item{run_1_name}{The name of the first run to compare.}

\item{run_2_name}{The name of the second run to compare.}
}
\value{
Called for the side effect of producing plots.
}
\description{
Generates \dQuote{CI} (credible interval) line plots that compare
the results of two global model runs. Country- and aggregate-level
(if \code{compare_aggregates} is \code{TRUE}) plots are available.
}
\seealso{
\code{\link{compare_runs_fishbone_plots}}.
}
\author{
Mark Wheldon (wrapper) based on underlying function by Jin
Rou New and Leontine Alkema.

Jin Rou New, Leontine Alkema, Mark Wheldon
}
