% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{do_global_all_women_run}
\alias{do_global_all_women_run}
\title{Do Complete Run of FPEM}
\usage{
do_global_all_women_run(
  run_desc = "",
  age_group = "15-49",
  estimation_iterations = 3,
  burn_in_iterations = 1,
  steps_before_progress_report = 4,
  thinning = 2,
  chain_nums = 1:3,
  run_in_parallel = TRUE,
  input_data_folder_path = system.file("extdata", package = "FPEMglobal"),
  data_csv_filename = paste0("data_cp_model_all_women_", age_group, ".csv"),
  region_information_csv_filename = "country_and_area_classification.csv",
  denominator_counts_csv_filename = paste0("number_of_women_", age_group, ".csv"),
  countries_for_aggregates_csv_filename = "countries_mwra_195.csv",
  countries_in_CI_plots_csv_filename = NULL,
  special_aggregates_name = NULL,
  start_year = 1970.5,
  end_year = 2030.5,
  years_change = matrix(c(1990.5, 2000.5, 2000.5, 2018.5, 2018.5, 2030.5, 2012.5,
    2018.5, 2012.5, 2020.5, 2012.5, 2017.5), ncol = 2, byrow = TRUE),
  years_change2 = matrix(c(2005.5, 2010.5, 2015.5, 2000.5, 2005.5, 2010.5, 1995.5,
    2000.5, 2005.5, 1990.5, 1995.5, 2000.5, 1990.5, 2000.5, 2010.5), ncol = 3, byrow =
    TRUE),
  plot_CI_changes_years = c(floor(start_year), floor(end_year)),
  make_any_aggregates = TRUE,
  adjust_medians = FALSE,
  run_name_override_married = NULL,
  run_name_override_unmarried = NULL,
  run_name_override_all_women = NULL,
  model_diagnostics = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{run_desc}{Character. Brief note to be appended to the auto-generated
\code{run_name}. Ignored if \code{run_name_override} is non-\code{NULL}.}

\item{age_group}{Character. The age group for which a run of the model is
desired, specified in the format \dQuote{xx-yy}, where \dQuote{xx} is
the start age, \dQuote{yy} is the end age, e.g., \code{"15-49"} or
\code{"15-19"}. This is used to form the run name if
\code{run_name_override} is \code{NULL}, to name of the file containing
prevalence data if \code{data_csv_filename} is \code{NULL}, and to
select rows from the denominator counts file if post-processing is done
(see \code{\link{post_process_mcmc}}). Only "15-49" is currently supported.}

\item{estimation_iterations}{Numeric. Number of MCMC iterations that should
be \emph{saved}. This is \emph{before} \code{thinning}.}

\item{burn_in_iterations}{Numeric. Number of MCMC iterations that should be
run as burn-in before starting to save them.}

\item{steps_before_progress_report}{Numeric. The number of times progress
should reported during MCMC sampling.}

\item{thinning}{Numeric. The actual number of iterations saved is
\eqn{\frac{\code{estimation_iterations}}{\code{thinning}}}{\code{estimation_iterations}/\code{thinning}}.}

\item{chain_nums}{Numeric. The number of MCMC chains to run, \emph{as a
sequence}. E.g., for three chains use \code{1:3}. You need to run at
least two chains for post-processing to be successful.}

\item{run_in_parallel}{Logical. Determines if MCMC chains are run in
parallel. Parallel running requires package
#ifdef windows
\pkg{doParallel}.
#endif
#ifndef windows

\pkg{doMC}.
#endif
Defaults to serial running if \code{run_in_parallel = TRUE} but the package
is not available.}

\item{input_data_folder_path}{File path to folder containing
\emph{all} input data. If
\code{NULL} the value of \code{data_csv_filename}, etc., will
be passed to \code{\link{file.path}} as-is. Otherwise,
\code{file.path(input_data_folder_path, data_csv_filename)}
will be passed. The default value points to the data directory
supplied with the package.}

\item{data_csv_filename}{Filename of the \file{.csv} file
containing country-level prevalence data. See
\dQuote{Details}.}

\item{region_information_csv_filename}{Filename of the \file{.csv}
file containing classifications of countries in sub-regions,
regions, etc. See \dQuote{Details}.}

\item{denominator_counts_csv_filename}{File path. Filepath to \file{.csv}
file with denominator counts (married and unmarried) for this
\code{age_group}. If \code{NULL}, defaults to \code{paste0("women_",
    \code{age_group}, ".csv")}.}

\item{countries_for_aggregates_csv_filename}{Name of the
\file{.csv} file listing countries that will be used in
constructing country aggregates.}

\item{countries_in_CI_plots_csv_filename}{Name of \file{.csv} file
that lists the countries to be included in the main
country-level indicator plots. These are the plots saved in
\file{\code{output_folder_path}/fig/\var{[run
    name]}CIs.pdf}. The format is the same as
\code{countries_for_aggregates_csv_filename}. The file is
looked for in \code{input_data_folder_path}. Countries appear
in the \file{.pdf} in the same order as they are listed in
\code{countries_in_CI_plots_csv_filename}.}

\item{special_aggregates_name}{Character vector of names
(\emph{not} filenames) of any speical aggregates
desired. There must be a corresponding file with name
\file{\code{special_aggregates_name}.csv} in
\code{input_data_folder_path} that defines the special
aggregates. See \dQuote{Details}.}

\item{start_year}{Estimates and projections are produced for a
specified time interval. This is the start year of that
interval.}

\item{end_year}{Estimates and projections are produced for a
specified time interval. This is the end year of that
interval.}

\item{years_change}{A two-column matrix giving the year pairs (as
rows) between which probabilistic estimates of changes in the
indicators are desired.}

\item{years_change2}{A three-column matrix giving the year triples
(as rows) among which to compute probabilistic estimates of
change-in-changes.}

\item{plot_CI_changes_years}{Vector of length two (if longer, only
the first and last elements are used). Declares the years to
be used to make the \dQuote{fish bone} plots, i.e., the plots
appearing in files
\file{\code{output_folder_path}/fig/\var{[run
    name]}_CIspropsubregional_modern_UNPD.pdf}. These must be in
the range of \code{start_year} and \code{end_year} passed to
\code{\link{post_process_mcmc}}.}

\item{adjust_medians}{Logical. Should adjusted medians outputs be
produced in addition to unadjusted outputs?}

\item{model_diagnostics}{Logical; should convergence diagnostics
and WAIC be computed?}

\item{verbose}{Logical; print lots and lots of messages about
progress?}
}
\value{
A \emph{list} of run names for married, unmarried, and all women
runs, returned invisibly. Results are saved to
\file{\code{output_folder_path}}.
}
\description{
Do married and unmarried runs of the global model, and produce
results. Then combine to produce all women results.
}
\details{
A \dQuote{run name} is assigned to each run of the model to identify it. The
run name is used to name the directory in which outputs are stored and the
filenames of certain results such as tables and plots. By default, it is
auto-generated by concatenating the date-time
(via \code{format(Sys.time(), "\%y\%m\%d_\%H\%M\%S")}),
marital group, age group and (optionally)
\code{run_desc}. If \code{run_desc} does not afford enough control,
\code{run_name_override} can be used to completely override
auto-generation. The function returns the run name as a character string.

See \code{system.file("extdata", "data_cp_model_all_women_15-49.csv")} for
how the prevalence data input should be formatted. Assume all columns are
required.

See \code{system.file("extdata", "country_and_area_classification.csv")} for
how the country classification file should be formatted. Assume all columns
are required.
}
\examples{
vignette("FPEMglobal")
}
\seealso{
\code{\link{do_global_run}} for just married or unmarried women runs.
}
\author{
Mark Wheldon, Andrew Tait
}
