% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{do_global_run}
\alias{do_global_run}
\title{Run the global FPEM model for married or unmarried women}
\usage{
do_global_run(
  marital_group = c("married", "unmarried"),
  age_group = "15-49",
  estimation_iterations = 3,
  burn_in_iterations = 1,
  number_incremental_backups = 4,
  thinning = 2,
  chain_nums = 1:3,
  set_seed_chains = 1,
  run_in_parallel = isTRUE(length(chain_nums) > 1),
  include_AR = TRUE,
  run_name = NULL,
  output_dir_path = file.path(getwd(), "output"),
  input_data_folder_path = system.file("extdata", package = "FPEMglobal"),
  data_csv_filename = paste0("data_cp_model_all_women_", age_group, ".csv"),
  region_information_csv_filename = "country_and_area_classification.csv",
  denominator_counts_csv_filename = paste0("number_of_women_", age_group, ".csv"),
  countries_for_aggregates_csv_filename = "countries_mwra_195.csv",
  countries_in_CI_plots_csv_filename = "countries_mwra_195.csv",
  special_aggregates_name = NULL,
  start_year = 1970.5,
  end_year = 2030.5,
  years_change = matrix(c(1990.5, 2000.5, 2000.5, 2018.5, 2018.5, 2030.5, 2012.5, 2018.5,
    2012.5, 2020.5, 2012.5, 2017.5), ncol = 2, byrow = TRUE),
  years_change2 = matrix(c(2005.5, 2010.5, 2015.5, 2000.5, 2005.5, 2010.5, 1995.5,
    2000.5, 2005.5, 1990.5, 1995.5, 2000.5, 1990.5, 2000.5, 2010.5), ncol = 3, byrow =
    TRUE),
  make_any_results = TRUE,
  plot_barchart_years = c(floor(start_year), floor(median(c(start_year, end_year))),
    floor(end_year)),
  plot_CI_changes_years = c(floor(start_year), floor(end_year)),
  make_all_bar_charts = TRUE,
  plot_maps_shapefile_folder = NULL,
  plot_maps_years = floor(median(c(start_year, end_year))),
  data_info_plot_years = c(1990, 2000, 2010),
  make_any_aggregates = TRUE,
  adjust_medians = TRUE,
  age_ratios_age_total_run_name = NULL,
  age_ratios_age_total_run_dir_path = NULL,
  age_ratios_age_total_denominator_counts_csv_filename = "number_of_women_15-49.csv",
  age_ratios_age_total_denominator_counts_folder_path = NULL,
  model_diagnostics = TRUE
)
}
\arguments{
\item{marital_group}{Character. The marital group for which a run of the
model is desired.}

\item{age_group}{Character. The age group for which a run of the model is
desired, specified in the format \dQuote{xx-yy}, where \dQuote{xx} is
the start age, \dQuote{yy} is the end age, e.g., \code{"15-49"} or
\code{"15-19"}. This is used to form the run name if
\code{run_name_override} is \code{NULL}, to name of the file containing
prevalence data if \code{data_csv_filename} is \code{NULL}, and to
select rows from the denominator counts file if post-processing is done
(see \code{\link{post_process_mcmc}}).}

\item{estimation_iterations}{Numeric. Number of MCMC iterations that should
be \emph{saved}. This is \emph{before} \code{thinning}.}

\item{burn_in_iterations}{Numeric. Number of MCMC iterations that should be
run as burn-in before starting to save them.}

\item{number_incremental_backups}{Numeric. The number of times MCMC chains
should be written to disc during the model run. These can be used to
restart a run after an interruption.}

\item{thinning}{Numeric. The actual number of iterations saved is
\eqn{\frac{\code{estimation_iterations}}{\code{thinning}}}{\code{estimation_iterations}/\code{thinning}}.}

\item{chain_nums}{Numeric. The number of MCMC chains to run, \emph{as a
sequence}. E.g., for three chains use \code{1:3}. You need to run at
least two chains for post-processing to be successful.}

\item{set_seed_chains}{Set the random seed passed to \code{JAGS}. For
multiple chains, the seed for each is \code{set_seed_chains} multiplied
by the chain number.}

\item{run_in_parallel}{Logical. Determines if MCMC chains are run in
parallel. Parallel running requires package
#ifdef windows
\pkg{doParallel}.
#endif
#ifndef windows

\pkg{doParallel} or \pkg{doMC}.
#endif
Defaults to serial running if \code{run_in_parallel = TRUE} but the package
is not available.}

\item{include_AR}{Logical; should the auto-regressive component of the model
be estimated. Used mainly for testing.}

\item{run_name}{Character. When \code{NULL} (default), this is set
automatically by concatenating the current date (yyyy-mm-dd),
\code{age_group} and \code{marital_group}. Alternatively, a user defined
run name to override default generation.}

\item{output_dir_path}{Path of the top-level directory in which to save
results. Sub-directories named \code{run_name} will be created inside
this directory for each run.}

\item{input_data_folder_path}{File path to folder containing \emph{all}
input data (except any map shapefiles). If \code{NULL} the value of
\code{data_csv_filename}, etc., will be passed to
\code{\link{file.path}} as-is. Otherwise,
\code{file.path(input_data_folder_path, data_csv_filename)} will be
passed. The default value points to the data directory supplied with the
package.}

\item{data_csv_filename}{Filename of the \file{.csv} file containing
country-level prevalence data. See \dQuote{Details}.}

\item{region_information_csv_filename}{Filename of the \file{.csv} file
containing classifications of countries in sub-regions, regions, etc.
See \dQuote{Details}.}

\item{denominator_counts_csv_filename}{Name of the \file{.csv}
file containing estimates and projections of the number of
women by marital status, age, and year. See \dQuote{Details}.}

\item{countries_for_aggregates_csv_filename}{Name of the
\file{.csv} file listing countries that will be used in
constructing country aggregates.}

\item{countries_in_CI_plots_csv_filename}{Name of \file{.csv} file
that lists the countries to be included in the main
country-level indicator plots. These are the plots saved in
\file{\code{run_dir_path}/fig/\var{[run
    name]}CIs.pdf}. The format is the same as
\code{countries_for_aggregates_csv_filename}. The file is
looked for in \code{input_data_folder_path}. Countries appear
in the \file{.pdf} in the same order as they are listed in
\code{countries_in_CI_plots_csv_filename}.}

\item{special_aggregates_name}{Character vector of names
(\emph{not} filenames) of any speical aggregates
desired. There must be a corresponding file with name
\file{\code{special_aggregates_name}.csv} in
\code{input_data_folder_path} that defines the special
aggregates. See \dQuote{Details}.}

\item{start_year}{Estimates and projections are produced for a
specified time interval. This is the start year of that
interval.}

\item{end_year}{Estimates and projections are produced for a
specified time interval. This is the end year of that
interval.}

\item{years_change}{A two-column matrix giving the year pairs (as
rows) between which probabilistic estimates of changes in the
indicators are desired.}

\item{years_change2}{A three-column matrix giving the year triples
(as rows) among which to compute probabilistic estimates of
change-in-changes.}

\item{make_any_results}{Logical. Should tables and plots be produced? If
\code{FALSE}, the arguments that pertain to specific plots or tables are
ineffective.}

\item{plot_barchart_years}{Vector of years for which bar charts
should be produced. These are saved to
\file{\code{run_dir_path}/fig/barchart}. These must be
in the range of \code{start_year} and \code{end_year} passed
to \code{\link{post_process_mcmc}}.}

\item{plot_CI_changes_years}{Vector of length two (if longer, only
the first and last elements are used). Declares the years to
be used to make the \dQuote{fish bone} plots, i.e., the plots
appearing in files
\file{\code{run_dir_path}/fig/\var{[run
    name]}_CIspropsubregional_modern_UNPD.pdf}. These must be in
the range of \code{start_year} and \code{end_year} passed to
\code{\link{post_process_mcmc}}.}

\item{make_all_bar_charts}{Logical. Produce barcharts? If
\code{NULL} but \code{plot_barchart_years} is non-\code{NULL},
is reset to \code{TRUE}.}

\item{plot_maps_shapefile_folder}{Path to directory containing
shapefiles for maps. Only needed if \code{plot_maps} is
\code{TRUE}. This is \emph{not} looked for in
\code{input_data_folder_path}; it should be a full path to the
shape files.}

\item{plot_maps_years}{Vector of years for which maps are desired;
the full set of maps are produced for each of the years
listed. These must be in the range of \code{start_year} and
\code{end_year} passed to \code{\link{post_process_mcmc}}.}

\item{make_any_aggregates}{Logical. Should country aggregates of
any kind (including default aggregates) be produced?}

\item{adjust_medians}{Logical. Should adjusted medians outputs be
produced in addition to unadjusted outputs?}

\item{age_ratios_age_total_run_name}{Run name of the 15--49 run to
use as the denominator for age ratios. Calculate ratios of
users in a subset age range (e.g., 15--19) to users in the
total age range (15--49) from this run. Requires a completed
15--49 run.}

\item{age_ratios_age_total_run_dir_path}{Alternative way of
specifying the run to use to make age ratios (see
\code{age_ratios_age_total_run_name}. File path to output
directory of the 15--49 run to use to make age ratios.}

\item{age_ratios_age_total_denominator_counts_csv_filename}{Name
of the \file{.csv} file containing estimates and projections
of the number of women by marital status, age, and year, for
the age group 15--49. Only used if \code{make_age_ratios} is
\code{TRUE}. Searched for in
\code{age_ratios_age_total_denominator_counts_folder_path}.}

\item{age_ratios_age_total_denominator_counts_folder_path}{Path to
\code{age_ratios_age_total_denominator_counts_csv_filename}. If
\code{NULL}, defaults to
\code{file.path(age_ratios_age_total_run_dir_path,
    "data")}.}

\item{model_diagnostics}{Logical; should convergence diagnostics
and WAIC be computed? These are not re-done if the folder
\sQuote{\code{run_dir_path}/convergence} exists.}
}
\value{
A name for the run, \code{run_name}, returned invisibly as a
character string.
}
\description{
This generates the MCMC chains using \code{\link{do_global_mcmc}},
post-processes them using \code{\link{post_process_mcmc}} and
produces plots and tables using \code{\link{make_results}}.
}
\details{
See \dQuote{Details} in the help file for \code{\link{do_global_all_women_run}}.
}
\examples{
vignette("FPEMglobal_Intro")

}
\references{
Kettunen, J. et al. (2012) Genome-wide association
study identifies multiple loci influencing human serum metabolite
levels. Nat Genet advance online publication.
\url{http://dx.doi.org/10.1038/ng.1073.}
}
\seealso{
\code{\link{combine_runs}} to create all women
results from married and unmarried women runs;
\code{\link{do_global_all_women_run}} to do married, unmarried,
\emph{and all women runs}, and produce results, all in one call.
}
\author{
Mark Wheldon, Andrew Tait
}
