% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface_utils.R
\name{get_run_name_from_args}
\alias{get_run_name_from_args}
\title{Get run_name from an argument list}
\usage{
get_run_name_from_args(args)
}
\value{
Run name as a character string.
}
\description{
Get run name from 'global_mcmc_args.RData',
'post_process_args.RData', or 'combine_runs_args.RData'.
}
\author{
Mark Wheldon
}
