% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface_utils.R
\name{make_global_summary_file_path}
\alias{make_global_summary_file_path}
\title{Construct path to global summary file}
\usage{
make_global_summary_file_path(output_folder_path = NULL, check = TRUE)
}
\arguments{
\item{output_folder_path}{Optional path to folder containing the file. By
default, only file filename will be returned.}

\item{check}{(Logical) Should a warning be given if the file does not exist?}
}
\value{
File path as a character string.
}
\description{
Defines the file name for, and constructs a file path pointing to, the
summary of a global run. This file is used as the input to a one-country
run. It contains the posterior means of hierarchical parameters.
}
\author{
Mark C Wheldon
}
