% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{make_results}
\alias{make_results}
\title{Summarize a post-processed global or one-country run of FPEM}
\usage{
make_results(
  run_name,
  output_folder_path = file.path("output", run_name),
  input_data_folder_path = NULL,
  denominator_counts_csv_filename = NULL,
  countries_for_aggregates_csv_filename = "countries_mwra_195.csv",
  region_information_csv_filename = "country_and_area_classification.csv",
  data_info_plot_years = c(1990, 2000, 2010),
  plot_CI_changes_years = NULL,
  countries_in_CI_plots_csv_filename = NULL,
  adjust_medians = FALSE,
  special_aggregates_name = NULL,
  all_women = NULL
)
}
\arguments{
\item{run_name}{The name of the run to post-process.}

\item{output_folder_path}{Filepath to directory where outputs
should be saved. If \code{NULL}, defaults to
\code{file.path("output", run_name)}.}

\item{input_data_folder_path}{File path to folder containing
\emph{all} input data. If
\code{NULL} the value of \code{data_csv_filename}, etc., will
be passed to \code{\link{file.path}} as-is. Otherwise,
\code{file.path(input_data_folder_path, data_csv_filename)}
will be passed. The default value points to the data directory
supplied with the package.}

\item{denominator_counts_csv_filename}{Name of the \file{.csv}
file containing estimates and projections of the number of
women by marital status, age, and year. See \dQuote{Details}.}

\item{countries_for_aggregates_csv_filename}{Name of the
\file{.csv} file listing countries that will be used in
constructing country aggregates.}

\item{region_information_csv_filename}{Filename of the \file{.csv}
file containing classifications of countries in sub-regions,
regions, etc. See \dQuote{Details}.}

\item{plot_CI_changes_years}{Vector of length two (if longer, only
the first and last elements are used). Declares the years to
be used to make the \dQuote{fish bone} plots, i.e., the plots
appearing in files
\file{\code{output_folder_path}/fig/\var{[run
    name]}_CIspropsubregional_modern_UNPD.pdf}. These must be in
the range of \code{start_year} and \code{end_year} passed to
\code{\link{post_process_mcmc}}.}

\item{countries_in_CI_plots_csv_filename}{Name of \file{.csv} file
that lists the countries to be included in the main
country-level indicator plots. These are the plots saved in
\file{\code{output_folder_path}/fig/\var{[run
    name]}CIs.pdf}. The format is the same as
\code{countries_for_aggregates_csv_filename}. The file is
looked for in \code{input_data_folder_path}. Countries appear
in the \file{.pdf} in the same order as they are listed in
\code{countries_in_CI_plots_csv_filename}.}

\item{adjust_medians}{Logical. Should adjusted medians outputs be
produced in addition to unadjusted outputs?}

\item{special_aggregates_name}{\emph{name} for special aggregates,
if any. The default, \code{NULL}, means no special aggregates
are produced. Note: this is \emph{not} a filename, although a
corresponding file named
\file{\code{special_aggregates_name}.csv} must be present in
\code{input_data_folder_path}.}

\item{all_women}{Logical; is the run an all women run such as the
kind produced by \code{\link{combine_runs}} or
\code{\link{do_global_all_women_run}}? If \code{NULL} an
attempt is made to determine this automatically from
\file{mcmc.meta.rda} in the output folder of the run.}

\item{married_women_run_name}{Run name of a married women
run.}

\item{unmarried_women_run_name}{Run name of a unmarried women
run. See \code{married_women_run_name}.}

\item{married_women_output_folder_path}{Path to directory
containing outputs for a married women run. See \code{married_women_run_name}.}

\item{unmarried_women_output_folder_path}{Path to directory
containing outputs for a unmarried women run. See \code{married_women_output_folder_path}.}
}
\value{
\code{run_name} invisibly as a character string.
}
\description{
Once post-processed, a global or one-country model can be used to generate
results in the form of summary figures and tables of key
indicators. Summaries are in the form of quantiles of posterior marginal
distributions, and means for some indicators.
}
\details{
The indicators for which these summaries are generated are, by
country and country aggregate, by year:
\itemize{
\item Contraceptive prevalence, any method (CP)
\item Contraceptive prevalence, traditional methods (CP modern)
\item Contraceptive prevalence, modern methods (CP traditional)
\item Unmet need for (modern) contraception (unmet need)
\item CP + Unmet need
\item Met demand for any method of contraception
\item Met demand for modern methods of contraception
\item Ratio of CP modern to CP
\item Probability that met demand for modern methods is greater than 75 percent
}
}
\examples{
vignette("FPEMglobal")

}
\seealso{
\code{\link{do_global_run}} (which calls this function)
to generate MCMC results for married or unmarried women,
post-process, and produce results all in one call;
\code{\link{combine_runs}} to create all women results
from married and unmarried women runs;
\code{\link{do_global_all_women_run}} to do married, unmarried,
\emph{and all women runs}, and produce results, all in one
call.
}
\author{
Mark Wheldon, Andrew Tait
}
