% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{pkg_files_included}
\alias{pkg_files_included}
\title{List of files included with package}
\usage{
pkg_files_included()
}
\value{
A nested list containing filenames and paths.
}
\description{
FPEMglobal comes bundled with all the input data files needed to
re-produce published model runs. This function will return a list
giving the filenames and paths to these files. They are grouped by
the package functions that take them as inputs. See
\dQuote{Description} for further details.
}
\details{
The files in the list are:
\description{
\item{model_run_inputs}{
\description{
\item{input_data_1549}{Raw input data for ages 15--49}
\item{region_information}{Regional and other aggregations of countries; used in model hierarchy.}
}
}
\item{post_process_inputs}{
\description{
\item{denominator_counts_1549}{Number of women by marital status, aged 15--49. These are used to convert proportions into counts, which are then used to construct the aggregates.}
\item{countries_for_aggregates}{Countries that are included in the set used to construct all aggregates. This set originally contained 195 countries but has grown slightly.}
}
}
\description{special_aggregates}{List of all files defining special aggregates, including SDG geographic regions and World Bank investment groups.}
}
}
\author{
Mark Wheldon
}
