% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{post_process_mcmc}
\alias{post_process_mcmc}
\title{Post process MCMC chains from a global run of FPEM}
\usage{
post_process_mcmc(
  run_name,
  output_folder_path = file.path("output", run_name),
  input_data_folder_path = NULL,
  denominator_counts_csv_filename = NULL,
  countries_for_aggregates_csv_filename = "countries_mwra_195.csv",
  start_year = 1970.5,
  end_year = 2030.5,
  years_change = matrix(c(1990.5, 2000.5, 2000.5, 2019.5, 2019.5, 2030.5, 2012.5,
    2019.5, 2012.5, 2019.5, 2012.5, 2020.5), ncol = 2, byrow = TRUE),
  years_change2 = matrix(c(2005.5, 2010.5, 2015.5, 2000.5, 2005.5, 2010.5, 1995.5,
    2000.5, 2005.5, 1990.5, 1995.5, 2000.5, 1990.5, 2000.5, 2010.5, 2000.5, 2010.5,
    2019.5), ncol = 3, byrow = TRUE),
  model_diagnostics = TRUE,
  special_aggregates_name = NULL,
  age_ratios_age_total_run_name = NULL,
  age_ratios_age_total_output_folder_path = NULL,
  age_ratios_age_total_denominator_counts_csv_filename = "number_of_women_15-49.csv",
  age_ratios_age_total_denominator_counts_folder_path = NULL,
  all_women = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The name of the run to post-process.}

\item{output_folder_path}{Filepath to directory where outputs
should be saved. If \code{NULL}, defaults to
\code{file.path("output", run_name)}.}

\item{input_data_folder_path}{File path to folder containing
\emph{all} input data (except any map shapefiles). If
\code{NULL} the value is taken from
\code{file.path(output_folder_path, "global_mcmc_args.RData")}
if that file exists, otherwise \code{file.path(output_folder_path, "data")}.}

\item{denominator_counts_csv_filename}{Name of the \file{.csv}
file containing estimates and projections of the number of
women by marital status, age, and year. See \dQuote{Details}.}

\item{countries_for_aggregates_csv_filename}{Name of the
\file{.csv} file listing countries that will be used in constructing country aggregates.}

\item{start_year}{Estimates and projections are produced for a
specified time interval. This is the start year of that
interval.}

\item{end_year}{Estimates and projections are produced for a
specified time interval. This is the end year of that
interval.}

\item{years_change}{A two-column matrix giving the year pairs (as
rows) between which probabilistic estimates of changes in the
indicators are desired.}

\item{years_change2}{A three-column matrix giving the year triples
(as rows) among which to compute probabilistic estimates of
change-in-changes.}

\item{model_diagnostics}{Logical; should convergence diagnostics
and WAIC be computed? These are not re-done if the folder \sQuote{\code{output_folder_path}/convergence} exists.}

\item{special_aggregates_name}{Character vector of names
(\emph{not} filenames) of any speical aggregates
desired. There must be a corresponding file with name
\file{\code{special_aggregates_name}.csv} in
\code{input_data_folder_path} that defines the special
aggregates. See \dQuote{Details}.}

\item{age_ratios_age_total_run_name}{Run name of the 15--49 run to use as the denominator for age ratios. Calculate ratios of users in a subset age range (e.g., 15--19) to users in the total age range (15--49) from this run. Requires a completed 15--49 run.}

\item{age_ratios_age_total_output_folder_path}{Alternative way of specifying the run to use to make age ratios (see \code{age_ratios_age_total_run_name}. File path to output
directory of the 15--49 run to use to make age ratios.}

\item{age_ratios_age_total_denominator_counts_csv_filename}{Name of
the \file{.csv} file containing estimates and projections of
the number of women by marital status, age, and year, for the
age group 15--49. Only used if \code{make_age_ratios} is
\code{TRUE}. Searched for in \code{age_ratios_age_total_denominator_counts_folder_path}.}

\item{age_ratios_age_total_denominator_counts_folder_path}{Path to
\code{age_ratios_age_total_denominator_counts_csv_filename}. If
\code{NULL}, defaults to
\code{file.path(age_ratios_age_total_output_folder_path, "data")}.}

\item{all_women}{Logical; is the run an all women run such as the
kind produced by \code{\link{combine_runs}} or
\code{\link{do_global_all_women_run}}? If \code{NULL} an
attempt is made to determine this automatically from
\file{mcmc.meta.rda} in the output folder of the run.}

\item{verbose}{Logical; print lots and lots of messages about
progress?}
}
\value{
The run name (invisibly). The function is mainly called
for its side effects.
}
\description{
MCMC chains from a global run of FPEM (via
\code{\link{do_global_mcmc}}) must be
post-processed before summary results (tables, plots) can be produced. This
function does the post-processing and saves the results in
\code{output_folder_path}. You need at least two chains for post-processing
to work.
}
\details{
The counts of women by marital status, age, and year in
\code{denominator_counts_csv_filename} are used to convert prevalence
proportions to counts of women by contraceptive use status. See
\code{system.file("extdata", "data_cp_model_all_women_15-49.csv", package = "FPEMglobal")}
for an example of how the file should be formatted. Assume all columns are
required.

By default, results are produced at the country level and for geographic
country aggregates. These are the sub-regions and regions of the \dfn{UN
M49} classification (\cite{UN DESA Statistics Division, 2017}).
For additional aggregates, argument
\code{special_aggregates_name} may be used. The argument takes the
\emph{name} of the aggregate. There must be a corresponding \file{.csv} file
in \code{input_data_folder_path}. The aggregate \dQuote{WHO_regions} is
included with the package; see \code{system.file("extdata", "WHO_regions.csv", package = "FPEMglobal")}
for the required format. Assum all columns are required.
}
\examples{
vignette("FPEMglobal_Intro")
}
\references{
UN DESA Statistics Division,
(2017) \emph{Standard Country or Area Codes for Statistical
Use (M49)}. United Nations, Department of Economic and Social
Affairs.  \url{https://unstats.un.org/unsd/methodology/m49/}
}
\seealso{
\code{\link{do_global_run}} (which calls this function)
to generate MCMC results for married or unmarried women,
post-process, and produce results all in one call;
\code{\link{combine_runs}} to create all women results
from married and unmarried women runs;
\code{\link{do_global_all_women_run}} to do married,
unmarried, \emph{and all women runs}, and produce results, all
in one call.
}
\author{
Mark Wheldon, Andrew Tait
}
