% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface_utils.R
\name{validate_extra_config}
\alias{validate_extra_config}
\title{Validate 'extra arguments' supplied via '...'}
\usage{
validate_extra_config(.extra_config)
}
\arguments{
\item{.extra_config}{The \code{...} as a list. Coercion to a list needs to
be done by the caller.}
}
\value{
The validated \code{...} as a list.
}
\description{
The main user interface functions (e.g., \code{\link{do_global_mcmc}}) have
a \code{...} argument. This is so that extra arguments for doing one-country
runs can be passed in and the \pkg{FPEMglobal} functions re-used, rather
than duplicated them in the \pkg{FPEMcountry} package.
}
\author{
Mark C Wheldon
}
