% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{add_global_mcmc}
\alias{add_global_mcmc}
\title{Add an MCMC chain to a global run of FPEM}
\usage{
add_global_mcmc(
  run_name,
  chain_nums = 2,
  output_folder_path = file.path("output", run_name),
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The name of the run to add a chain to.}

\item{chain_nums}{number identifying chains to add. Unlike
\code{\link{do_global_mcmc}}, this can be a scalar. It must \emph{not}
be the number identifying a chain already created. See \dQuote{Details}.}

\item{output_folder_path}{Filepath to directory where outputs
should be saved. If \code{NULL}, defaults to
\code{file.path("output", run_name)}.}

\item{verbose}{Logical; print lots and lots of messages about
progress?}
}
\value{
A name for the run is returned invisibly as a character string. MCMC chains are
saved to \file{output_folder_path/temp.JAGSobjects}. They need to
be post-processed with \code{\link{post_process_mcmc}}. The run name
must be passed to \code{\link{post_process_mcmc}} to locate find the
saved chains for processing. Run names for married and unmarried runs
must also be passed to \code{\link{combine_runs}} to generate all
women MCMC results.
}
\description{
At least two chains are required to post-process the results of a run of the
FPEM model. Multiple chains can be run by \code{\link{do_global_mcmc}} but if additional chains are desired
this function can be used to add them. There must be at least one chain
already started (identified via \code{run_name}).
}
\examples{
vignette("FPEMglobal")
}
\author{
Mark Wheldon, Andrew Tait
}
