% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPEMglobal.R
\docType{package}
\name{FPEMglobal}
\alias{FPEMglobal}
\title{FPEMglobal: Contraceptive use package that provides estimates for all women. See vignette 'FPEMglobal-vignette' for explanation of use.}
\description{
FPEMglobal: Contraceptive use package that provides estimates for all women. See vignette 'FPEMglobal-vignette' for explanation of use.
}
