% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{make_results}
\alias{make_results}
\title{Summarize a post-processed global run of FPEM}
\usage{
make_results(
  run_name = NULL,
  output_folder_path = file.path("output", run_name),
  input_data_folder_path = NULL,
  countries_in_CI_plots_csv_filename = "countries_mwra_195.csv",
  CI_plots_years = NULL,
  plot_diagnostic_CI_plots = FALSE,
  make_all_bar_charts = NULL,
  plot_barchart_years = NULL,
  data_info_plot_years = c(1990, 2000, 2010),
  plot_CI_changes_years = NULL,
  plot_maps_shapefile_folder = NULL,
  plot_maps_years = NULL,
  plot_prior_post = NULL,
  plot_parameters = NULL,
  tabulate_changes = TRUE,
  adjust_medians = TRUE,
  make_any_aggregates = TRUE,
  special_aggregates_name = NULL,
  make_age_ratios = NULL,
  validation_keep_all = TRUE,
  validation_return_res_as_df = FALSE,
  all_women = NULL,
  verbose = FALSE
)
}
\arguments{
\item{run_name}{The name of the run to post-process.}

\item{output_folder_path}{Filepath to directory where outputs
should be saved. If \code{NULL}, defaults to
\code{file.path("output", run_name)}.}

\item{input_data_folder_path}{File path to folder containing
\emph{all} input data (except any map shapefiles). If
\code{NULL} the value of \code{data_csv_filename}, etc., will
be passed to \code{\link{file.path}} as-is. Otherwise,
\code{file.path(input_data_folder_path, data_csv_filename)}
will be passed. The default value points to the data directory
supplied with the package.}

\item{countries_in_CI_plots_csv_filename}{Name of \file{.csv} file
that lists the countries to be included in the main
country-level indicator plots. These are the plots saved in
\file{\code{output_folder_path}/fig/\var{[run
    name]}CIs.pdf}. The format is the same as
\code{countries_for_aggregates_csv_filename}. The file is
looked for in \code{input_data_folder_path}. Countries appear
in the \file{.pdf} in the same order as they are listed in
\code{countries_in_CI_plots_csv_filename}.}

\item{plot_diagnostic_CI_plots}{Logical. Produce diagnostic
versions of the main country-level indicator plots. These have
zoomed y-axes.}

\item{make_all_bar_charts}{Logical. Produce barcharts? If
\code{NULL} but \code{plot_barchart_years} is non-\code{NULL},
is reset to \code{TRUE}.}

\item{plot_barchart_years}{Vector of years for which bar charts
should be produced. These are saved to
\file{\code{output_folder_path}/fig/barchart}. These must be
in the range of \code{start_year} and \code{end_year} passed
to \code{\link{post_process_mcmc}}.}

\item{plot_CI_changes_years}{Vector of length two (if longer, only
the first and last elements are used). Declares the years to
be used to make the \dQuote{fish bone} plots, i.e., the plots
appearing in files
\file{\code{output_folder_path}/fig/\var{[run
    name]}_CIspropsubregional_modern_UNPD.pdf}. These must be in
the range of \code{start_year} and \code{end_year} passed to
\code{\link{post_process_mcmc}}.}

\item{plot_maps_shapefile_folder}{Path to directory containing
shapefiles for maps. Only needed if \code{plot_maps} is
\code{TRUE}. This is \emph{not} looked for in
\code{input_data_folder_path}; it should be a full path to the
shape files.}

\item{plot_maps_years}{Vector of years for which maps are desired;
the full set of maps are produced for each of the years
listed. These must be in the range of \code{start_year} and
\code{end_year} passed to \code{\link{post_process_mcmc}}.}

\item{adjust_medians}{Logical. Should adjusted medians outputs be
produced in addition to unadjusted outputs?}

\item{make_any_aggregates}{Logical. Should country aggregates of
any kind (including default aggregates) be produced?}

\item{special_aggregates_name}{\emph{name} for special aggregates,
if any. The default, \code{NULL}, means no special aggregates
are produced. Note: this is \emph{not} a filename, although a
corresponding file named
\file{\code{special_aggregates_name}.csv} must be present in
\code{input_data_folder_path}.}

\item{make_age_ratios}{Logical. Should results for age-ratios be
made? If \code{NULL} this is set based on the presence of
files \dQuote{res.country.age.ratio.rda} or
\dQuote{res.country.all.women.age.ratio.rda} in
\code{output_folder_path}.}

\item{validation_keep_all}{Logical. Should validation results be
saved? If \code{TRUE}, a list with elements \code{CI.df} and
\code{Error.df} is saved to
\file{\code{output_folder_path}/validn_repts/validation.results.rda}.}

\item{validation_return_res_as_df}{Logical; should the validation
results be returned as a data frame? If \code{FALSE} the
function returns the \code{run_name} invisibly as a character
string.}

\item{verbose}{Logical; print lots and lots of messages about
progress?}

\item{married_women_run_name}{Run name of a married women
run. Relevant if \code{\link{make_results}} is being run on an
all women run and age ratios for country aggregates are
wanted. If this argument is \code{NULL} (default), and
\code{all_women} is \code{TRUE} (explicitly or implicitly), an
attempt is made to find this folder by substituting
\dQuote{all_women} with \dQuote{married} in the
\code{run_name}.}

\item{unmarried_women_run_name}{Run name of a unmarried women
run. See \code{married_women_run_name}.}

\item{married_women_output_folder_path}{Path to directory
containing outputs for a married women run. See
\code{married_women_run_name}.}

\item{unmarried_women_output_folder_path}{Path to directory
containing outputs for a unmarried women run. See
\code{married_women_output_folder_path}.}
}
\value{
Either \code{run_name} invisibly as a character string or,
if \code{isTRUE(validation_return_res_as_df)}, a data frame
(non-invisibly) containing validation results.
}
\description{
Once post-processed, a global model can be used to generate
results in the form of summary figures and tables of key
indicators. Summaries are in the form of quantiles of posterior marginal
distributions, and means for some indicators.
}
\details{
The indicators for which these summaries are generated are, by
country and country aggregate, by year:
\itemize{
\item Contraceptive prevalence, any method (CP)
\item Contraceptive prevalence, traditional methods (CP modern)
\item Contraceptive prevalence, modern methods (CP traditional)
\item Unmet need for (modern) contraception (unmet need)
\item CP + Unmet need
\item Met demand for any method of contraception
\item Met demand for modern methods of contraception
\item Ratio of CP modern to CP
\item Probability that met demand for modern methods is greater than 75 percent
}
}
\section{Age ratios for aggregates}{

\code{\link{post_process_mcmc}} generates trajectories for the
age ratios for countries for married, unmarried, and all
women. However, \code{\link{combine_runs}}, not
\code{\link{post_process_mcmc}}, produces the trajectories for
age ratios for country aggregates, including those for married
and unmarried women. When \code{\link{make_results}} is
applied to an all women run, an attempt is made to find the
corresponding
}

\examples{
vignette("FPEMglobal_Intro")

}
\seealso{
\code{\link{do_global_run}} (which calls this function)
to generate MCMC results for married or unmarried women,
post-process, and produce results all in one call;
\code{\link{combine_runs}} to create all women results
from married and unmarried women runs;
\code{\link{do_global_all_women_run}} to do married, unmarried,
\emph{and all women runs}, and produce results, all in one
call.
}
\author{
Mark Wheldon, Andrew Tait
}
