% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{rename_global_run}
\alias{rename_global_run}
\title{Rename global run}
\usage{
rename_global_run(
  run_name = NULL,
  new_run_name,
  output_dir_path = file.path(getwd(), "output"),
  rename_output_folder = missing(run_dir_path),
  ignore = "^data$",
  verbose = deprecated(),
  output_folder_path = deprecated()
)
}
\arguments{
\item{run_name}{Character. Name of run to be renamed. If this is \code{NULL}
(default), the run name will be taken from the
\filename{global_mcmc_args.RData} file in \code{run_dir_path}.}

\item{new_run_name}{Character. New run name.}

\item{output_dir_path}{Path of the top-level directory in which to save
results. Sub-directories named \code{run_name} will be created inside
this directory for each run.}

\item{rename_output_folder}{Logical. Should an attempt be made to rename the
output folder? See \dQuote{Details}.}

\item{ignore}{Regular expression. Files and directories with names That
match will not be renamed.}

\item{run_dir_path}{Filepath to directory where outputs are saved. If
\code{NULL}, defaults to \code{file.path(output_dir_path, run_name)}.}
}
\value{
Called for side effect only.
}
\description{
Renames the files and directories , and meta data, of a global run of FPEM
for a single marital group.
}
\details{
If \code{rename_output_folder = TRUE}, an attempt is made to rename the
output folder. All matches of \code{run_name} are replaced with
\code{new_run_name}.

If found, the files \filename{global_mcmc_args.RData},
\filename{post_process_args.RData}, and \filename{combine_runs_args.RData},
all in \code{run_dir_path} are modified as follows:

\enumerate{
\item The element \code{run_name} is replaced with the \code{new_run_name}.
\item Element \code{renamed} is set to \code{TRUE} (it is created if it doesn't already exist).
\item Element \code{rename_list}, a character vector, is augmented by adding
the old run name (as per \code{run_name}) to the front as the first element.
}
}
\section{Note}{
 If you modify \code{ignore} make sure
\file{\code{run_dir_path}/data} still matches; it is not
a good idea to rename files in that directory.
}

\author{
Mark Wheldon
}
